+++ plr_exercise/models/__init__.py
@@ -1,0 +1,1 @@
+from .cnn import Net
